src/python/requirements.txt
pandas==2.0.3
requests==2.31.0
rpy2==3.5.15
python-dotenv==1.0.0
pyyaml==6.0
schedule==1.2.0